+++ payroll/src/main/java/com/playground/payroll/service/tax/dto/TaxDTO.java
@@ -1,6 +1,7 @@
 package com.playground.payroll.service.tax.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * DTO containing a tax table record.
@@ -13,10 +14,10 @@ public class TaxDTO implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
 	private Long id;	
-	private Integer salaryFrom;	
-	private Integer baseSalary;	
-	private Integer baseSalaryTax;	
-	private Double additionalRate;
+	private BigDecimal salaryFrom;	
+	private BigDecimal baseSalary;	
+	private BigDecimal baseSalaryTax;	
+	private BigDecimal additionalRate;
 	
 	public Long getId() {
 		return id;
@@ -24,28 +25,28 @@ public class TaxDTO implements Serializable {
 	public void setId(Long id) {
 		this.id = id;
 	}
-	public Integer getSalaryFrom() {
+	public BigDecimal getSalaryFrom() {
 		return salaryFrom;
 	}
-	public void setSalaryFrom(Integer salaryFrom) {
+	public void setSalaryFrom(BigDecimal salaryFrom) {
 		this.salaryFrom = salaryFrom;
 	}
-	public Integer getBaseSalary() {
+	public BigDecimal getBaseSalary() {
 		return baseSalary;
 	}
-	public void setBaseSalary(Integer baseSalary) {
+	public void setBaseSalary(BigDecimal baseSalary) {
 		this.baseSalary = baseSalary;
 	}
-	public Integer getBaseSalaryTax() {
+	public BigDecimal getBaseSalaryTax() {
 		return baseSalaryTax;
 	}
-	public void setBaseSalaryTax(Integer baseSalaryTax) {
+	public void setBaseSalaryTax(BigDecimal baseSalaryTax) {
 		this.baseSalaryTax = baseSalaryTax;
 	}
-	public Double getAdditionalRate() {
+	public BigDecimal getAdditionalRate() {
 		return additionalRate;
 	}
-	public void setAdditionalRate(Double additionalRate) {
+	public void setAdditionalRate(BigDecimal additionalRate) {
 		this.additionalRate = additionalRate;
 	}
